Module Doctree.Labels

val disambiguate_page : enter_subpages:bool -> Types.Page.t -> Types.Page.t

Colliding labels are allowed in the model but don't make sense in generators because we need to link to everything (eg. the TOC). Post-process the doctree, add a "_N" suffix to dupplicates, the first occurence is unchanged. Iterate through subpages.