Module Url.Path

type kind = [

| `Module

| `Page

| `LeafPage

| `ModuleType

| `Parameter of int

| `Class

| `ClassType

| `File

| `SourcePage

]

val pp_kind : Format.formatter -> kind -> unit

val string_of_kind : kind -> string

type t = {

kind : kind;

parent : t option;

name : string;

}

type any_pv = [

| Odoc_model.Paths.Identifier.Page.t_pv

| Odoc_model.Paths.Identifier.Signature.t_pv

| Odoc_model.Paths.Identifier.ClassSignature.t_pv

| Odoc_model.Paths.Identifier.SourcePage.t_pv

| Odoc_model.Paths.Identifier.SourceDir.t_pv

| Odoc_model.Paths.Identifier.AssetFile.t_pv

]

and any = any_pv Odoc_model.Paths.Identifier.id

val from_identifier : [< any_pv ] Odoc_model.Paths.Identifier.id -> t

val to_list : t -> (kind * string) list

val of_list : (kind * string) list -> t option

val split : is_dir:(kind -> [ `Always | `Never | `IfNotLast ]) -> (kind * string) list -> (kind * string) list * (kind * string) list

split is_dir path splits the list path into a directory and filename, based on the is_dir function. The function is_dir should return whether or not the path element kind should be a directory or not. If the function is_dir returns `IfNotLast then the element will be a directory only if it is not the last element in the path. The return value is a tuple of directory-type elements and filename-type elements. If the is_dir function can return `Always, the caller must be prepared to handle the case where the filename part is empty.